+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Xcode 8.3 or later and Swift 3.1 or later are required to build.  
   [Norio Nomura](https://github.com/norio-nomura)
+  
+* Nested configurations will now be merged with parent configurations, rather than replace them.  
+  [Stéphane Copin](https://github.com/stephanecopin/)
 
 ##### Enhancements
 
@@ -33,6 +36,10 @@
 * Improve performance when linting and correcting on Linux, matching behavior.  
   [JP Simard](https://github.com/jpsim)
   [#1577](https://github.com/realm/SwiftLint/issues/1577)
+  
+* Add support for merging nested configurations with parent configurations.  
+  [Stéphane Copin](https://github.com/stephanecopin/)
+  [#676](https://github.com/realm/SwiftLint/issues/676)
 
 ##### Bug Fixes